# Improved performance of System.Reflection.Emit API