… src/System.Private.CoreLib/src/System/Reflection/Emit/SignatureHelper.cs