# Performance Improvements to System.Reflection.Emit.ILGenerator

When researching how dynamically emitting IL works in the CoreClr I stumbled upon a couple blocks of code that looked like they could be changed to improve performance.

This investigation led me to make significant performance improvements to the ILGenerator class.

Below is an image of the benchmarks i ran on a couple of publicly facing routines of the ILGenerator.


|  Method  |  Cycles  |  Old   |  New  | Change  |
|---|:---:|:---:|:---:|:---:|
|  `ILGenerator.Emit(OpCode, long)`  |  128|  *2.031 us*  |  *961.8 ns*  |  **+111%**  |
|  `ILGenerator.Emit(OpCode)` |  128  |  *1.259 us*  |  *642.2 ns*  |  **+96%**  |
|  `ILGenerator.Emit(OpCode, double)` |  128 |  *1.995 us* |  *953.2 ns*  |  **+109%**  |
|  `ILGenerator.Emit(OpCode, int)` |  128  |  *1.625 us* |  *901.0 ns*  |  **+80%**  |
|  `ILGenerator.Emit(OpCode, byte)` |  128  | *1.383 us* | *769.6 ns* |  **+79%**  |
|  `ILGenerator.Emit(OpCode, float)` |  128  | *1.482 us* | *907.9 ns*  |  **+63%**  |


[Above table as Image of BenchmarkDotNet output (Outdated)](https://cdn.discordapp.com/attachments/406551726896709632/565117951729074187/aarst.PNG)

As you can see, a nearly 2x improvement across the board!
These improvements are not just limited to the routines in the benchmarks. Most, if not all ILGenerator.Emit routines should see some sort of improvement in performance.